[MISC] split type_(pack|list) pack algorithm detail header
